Update the ssatab phase name in me_ssa_tab.h
The me_ssa_tab.h use the ssatab phase's name as 'ssaTab'.
It is different with the maple_driver/defs/phases.def,
which is used as 'ssatab'. All the phases name in the phases.def
all write in lowercase, so I update the 'ssaTab' in me_ssa_tab.h
to 'ssatab'.